Move the advanced settings filter into JetpackFieldControls
